Create implement_ternary_search.dart
This Dart program defines a function ternarySearch that implements the ternary search algorithm as described in your pseudocode. The main function demonstrates how to use ternarySearch by searching for the value 4 in the list [1, 2, 3, 4, 5]. The expected output is "Element found at index: 3" since the value 4 is located at the index 3 in the list (considering 0-based indexing).